- playblack fixes (ClientLogin)
