+++ changelog.txt
@@ -1,6 +1,7 @@
 - known issue: Problem activating a second instance with the same account #32
 - known issue: Play from here in videos causes endless "opening stream" + error #29
 
+- playblack fixes (ClientLogin)
 - better seeking support
 Version 0.6.31
 - fix error issue around resume points
