+++ resources/lib/gdrive_api2.py
@@ -1541,55 +1541,6 @@ class gdrive(cloudservice):
         return mediaURLs
 
 
-    def getMediaSelection(self, mediaURLs, folderID, filename):
-
-        options = []
-        mediaURLs = sorted(mediaURLs)
-        for mediaURL in mediaURLs:
-            options.append(mediaURL.qualityDesc)
-            if mediaURL.qualityDesc == 'original':
-                originalURL = mediaURL.url
-
-        mediaURL = ''
-        if self.settings.playOriginal:
-            mediaURL = mediaurl.mediaurl(originalURL +'|' + self.getHeadersEncoded(self.useWRITELY), 'original', 0, 9999)
-            return mediaURL
-
-        #playbackPath = str(self.settings.cachePath) + '/' + str(filename) + '/'
-        (localResolutions,localFiles) = self.cache.getFiles(self)
-        totalList = localFiles + mediaURLs
-        mediaCount = len(localFiles)
-
-        if self.settings.promptQuality:
-            ret = xbmcgui.Dialog().select(self.addon.getLocalizedString(30033), localResolutions + options)
-            if ret > mediaCount:
-                mediaURL = totalList[ret]
-                mediaURL.url = totalList[ret].url
-            else:
-                mediaURL = mediaurl.mediaurl(str(totalList[ret]), 'offline', 0, 0)
-                mediaURL.offline = True
-
-        else:
-            if len(localFiles) == 0:
-                mediaURL = totalList[0]
-                mediaURL.url = totalList[0].url
-            else:
-                mediaURL = mediaurl.mediaurl(str(totalList[0]), 'offline', 0, 0)
-                mediaURL.offline = True
-
-
-#        elif self.settings.promptQuality and len(options) > 1 and not self.settings.cache:
-#            ret = xbmcgui.Dialog().select(self.addon.getLocalizedString(30033), options)
-#            mediaURL = mediaURLs[ret]
-#            if not self.settings.download:
-#                mediaURLs[ret].url = mediaURLs[ret].url +'|' + self.getHeadersEncoded(self.useWRITELY)
-
-#        else:
-#            mediaURLs[0].url = mediaURLs[0].url +'|' + self.getHeadersEncoded(self.useWRITELY)
-#            mediaURL = mediaURLs[0]
-
-        return mediaURL
-
 
 
     def setProperty(self, docid, key, value):
